Make Folder a facade.
